implement the pet card (#71)
* Implement PetCard component.

Make and style the PetCard component. The home
page just links to the petcard for testing. This
will be changed back to normal before merging.

* Update the PetCard with props

Add props to the PetCard to pass in user entered
information. The favorite button is not yet
functional.

* Update PetCard.css with Gio's changes

Update PetCard.css to be more responsive to
different image sizes.